+++ include/finufft/spreadinterp.h
@@ -7,6 +7,7 @@
 #ifndef SPREADINTERP_H
 #define SPREADINTERP_H
 
+#include <finufft/finufft_core.h>
 #include <finufft_spread_opts.h>
 
 /* Bitwise debugging timing flag (TF) defs; see finufft_spread_opts.flags.
